feat: move `ollamaApi.ts` to `src/client` for use by adpative-palette
